size_tは符号なし整数なので、負の値に減算できないため修正

+++ src/is_valid_path_utils.c
@@ -6,7 +6,7 @@
 /*   By: hkumagai <hkumagai@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/21 15:17:55 by hkumagai          #+#    #+#             */
-/*   Updated: 2022/10/25 14:04:09 by hkumagai         ###   ########.fr       */
+/*   Updated: 2022/10/26 15:08:37 by hkumagai         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	move_up(t_map *map, size_t x, size_t y)
 {
 	size_t	move_after_x;
-	size_t	move_after_y;
+	ssize_t	move_after_y;
 
 	move_after_x = x;
 	move_after_y = y - 1;
@@ -23,7 +23,7 @@ void	move_up(t_map *map, size_t x, size_t y)
 	{
 		if (map->map[move_after_y][move_after_x] != '1' && \
 			map->map[move_after_y][move_after_x] != '3')
-			move(map, move_after_x, move_after_y);
+			move(map, move_after_x, (size_t)move_after_y);
 	}
 }
 
@@ -59,7 +59,7 @@ void	move_right(t_map *map, size_t x, size_t y)
 
 void	move_left(t_map *map, size_t x, size_t y)
 {
-	size_t	move_after_x;
+	ssize_t	move_after_x;
 	size_t	move_after_y;
 
 	move_after_x = x - 1;
@@ -68,6 +68,6 @@ void	move_left(t_map *map, size_t x, size_t y)
 	{
 		if (map->map[move_after_y][move_after_x] != '1' && \
 			map->map[move_after_y][move_after_x] != '3')
-			move(map, move_after_x, move_after_y);
+			move(map, (size_t)move_after_x, move_after_y);
 	}
 }
